refactor middlwares
